refactor : battleStore, battlePhase, teamColors를 features/battle로 이동
