+++ frontend/src/features/battle/types/teamColors.ts
