+++ frontend/src/features/battle/utils/battlePhase.ts
