+++ frontend/src/pages/battlePage/components/discussion/test/DiscussionInput.test.tsx
@@ -12,7 +12,7 @@ let mockBattleProgress: {
 };
 let mockSelectedTeam: 'A' | 'B' = 'A';
 
-vi.mock('@/pages/battlePage/stores/battleStore', () => ({
+vi.mock('@/features/battle/stores/battleStore', () => ({
   useBattleStore: vi.fn((selector) => {
     const state = {
       battleProgress: mockBattleProgress,
