+++ frontend/src/pages/tutorialPage/battle/hooks/useTutorialBattleSetup.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import type { BattleInfo } from '@/commons/types/battle';
-import { useBattleStore } from '@/pages/battlePage/stores/battleStore';
+import { useBattleStore } from '@/features/battle/stores/battleStore';
 import { useAuthStore } from '@/commons/stores/authStore';
 import { createMockSocket } from '@/pages/tutorialPage/utils/mockSocket';
 import {
